Add a new Scene implementation for building the game worlds
Add a new Scene implementation for building the game worlds. This
is the next commit of the scene refactoring and continues the work
that was started in 5d9ad56ab24944491df1450dd904a3da7840101e

The SceneClass is a class used to build the "static" representation
of the scene. Static here means something that is pre-built before
the actual game launches most likely using a tool such as the Editor
application. The SceneNodeClass objects define the initial state of
the scene in terms of types of entities and their initial state and
placement in the scene. Once a new Scene instance is created using
the SceneClass object the SceneNodeClass objects are used to create
the Entities in the scene. During the game play the Scene can then
be dynamically changed by adding more entities to it or by removing
some of the existing entities.

The API for the Scene class is not yet complete and all of the
implementation for SceneClass is completely unoptimized.