Add an explicit call to QWindow::destroy when disposing of GfxWidget
Add an explicit call to QWindow::destroy when disposing of GfxWidget.

There seems to be an issue where the rendering gets slower and slower
until at some point the X server crashes and the X11 session is lost.
It seems that the underlying issue is about some GPU resources being
leaked.

Adding an explicit call to QWindow::destroy and hoping that it'd resolve
or mitigate this issue.

The nvidia-settings application can be used to quickly inspect the
system's overall GPU memory consumption.