Add a workaround for a Qt memory leak bug on Xorg and NVIDIA
Add a workaround for a Qt memory leak bug on Xorg and NVIDIA.

Basically there's a Qt bug that fails to clean up some OpenGL
resources which are shared with the X server when then parent
of the widget created by a call to QWidget::createWindowContainer
is set to nullptr. This leaves those resources still open in the X
server causing an increase in VRAM consumption that grows over time
as windows are opened and closed eventually leading to trashing
and X server failure.

More discussion can be found below:

"Xorg VRAM leak because of Qt/OpenGL Application"
https://forum.qt.io/topic/92179/xorg-vram-leak-because-of-qt-opengl-application/12

"Xorg VRAM leak because of Qt/OpenGL Application"
https://community.khronos.org/t/xorg-vram-leak-because-of-qt-opengl-application/76910/2

"NULL parenting a parent of a window container containing a QOpenGLWindow make Xorg leaks VRAM"
https://bugreports.qt.io/browse/QTBUG-69429