Change the semantics of MouseTool MouseRelease return value
Change the semantics of MouseTool MouseRelease return value. The
new semantics use the boolean return value to indicate whether
the tool is done or not.

Use the new semantics of MouseTool MouseRelease to allow the entity
placement tool to continue operation until explicitly cancelled
through pressing Esc. This makes it convenient to place multiple
objects rapidly.