Allow render tree traversal to finish early by setting a flag
Allow render tree traversal to finish early by setting a flag. For
example when looking for a specific node in the render tree in
order to discover its complete node-to-world transformation matrix
once the node is found there's no need to traverse the rest of the
render tree nodes. Instead a flag can be set to indicate early
completion of the traversal algorithm.