Improve the efficiency of some of the render tree operations
Improve the efficiency of some of the render tree operations. The
operations that start with a *known* child node that exists in the
render tree can now be optimized to with more efficient tree
traversal. I.e. to find the path from some child to route and to
it's now possible to use the RenderTree<T>::GetParent function.