Add first version of scripting variable support
Add first version of scripting variable support. Scripting variables
represent variables (members) associated with some "klass" type such
as an EntityClass. I.e any user defined entity class type may have a
number of variables (and constants) defined as part of the type. The
variables will then be exposed and accessible in the scripting through
some interface in the scripting environment. Variables can also be
marked as "read-only" effectively making them constants in the
scripting environment.

Note that the semantics of "constness" are such that we cannot really
represent that with cpp constness. const ScriptVar semantically means
that the variable cannot change it's semantics, i.e. type, name and
flags while its actual value that it holds can change. Therefore its
possible to set a new value to a const ScriptVar. The constness of the
value is also a runtime checked property.