Only consider the currently active tab for updates and rendering
Only consider the currently active tab for updates and rendering in
the main tab. There's no point to update/render other tabs since
they're not visible. However extra care must be taken to make sure
that when the rendering is skipped we're still syncing to vblank
(when it's set ON) on *some* surface. This is now achieved through
some more global variables in GfxWidget implementantion that help
us track whether vsync *did* happen on the current frame loop
iteration already. If it did not happen (i.e. the current widget
is *not* the one with vsynced surface) we look for the correct
surface from the global set of surfaces and perform an extra
buffer swap just for the sake of sync.