Add angular and linear velocity to the entity rigid body
Add angular and linear velocity to the entity rigid body. These
parameters double as initial values for the physics simulation and
then change behaviour depending on the type of the physics body.
- With kinematic bodies the velocites can be updated by the for
  example the animation system and the changes are reflected into
  to the physics simulation.
- With dynamic bodies the physics engine will update the body with
  the current instantaneous velocities.

For driving kinematic bodies in the physics world changing these
velocities should be the primary means of enabling motion.