Add a material dialog for visually choosing a material
Add a material dialog for visually choosing a material. This dialog
renders some visual representation of each and every material currently
in the workspace and then allows a material to be selected by clicking
on of the illustrations.

Add UI functionality for opening the material chooser dialog in the
entity widget.