Add first implementation and support for Lua based games
Add first implementation and support for Lua based games. There's
a new interface called "Game" which abstracts away the actual game
logic and the implementation of that logic. The game engine will
create an instance of Game and essentially treat it as one of the
subsystems (rendering, physics etc). The idea is that this interface
will be adequate for keeping the internals of the engine (keeping
time, servicing subsystems etc) separate from the actual game logic.

There's a simple barebones getting started implementation of the
game interface for Lua based games. This implementation will simply
forward the calls to the Lua script.