Add type aliases for the simple types used by gamelib for non gfx code
Add type aliases for the simple types used by gamelib for non gfx code.
The code that is not graphics related should not depend on types
defined in graphics/. Using a type alias is a step forward in terms
of removing that dependency. Ultimately the types should probably
move into base/ and then be used by graphics and gamelib etc.