Add game logical viewport setting and manipulation
Add game logical viewport setting and manipulation. This is the
logical viewport that defines the view into the game world. It can
be though of as a window into the game world and it has some position
and some dimension. The game can fully control both aspects of the
viewport through the Lua API. When the engine is about the render the
world it'll ask the game for its current logical game viewport. The
information is then used to map some region of the rendering surface
for the output to display the part of the game world that is within
the game's viewport. This mapping is completely agnostic regarding
the actual resolution of the viewing device (window resolution) which
then can be resized arbitrarily without changing which part of the
game gets shown. This design should help keep the actual games agnostic
about any actual display resolutions.