Track focus in/out events in GfxWidget and draw focus rect manually
Track focus in/out events in GfxWidget and draw focus rect manually.
Currently there's no visual indicator when the QWindow has the keyboard
focus or not. Internally we got a QWindow inside a QWidget in a container
so the drawing is not based on QWidget. That means that this functionality
doesn't work out of the box as it does for custom widgets that
are based on QWidget instead. We can draw the focus rect also self
but figuring out the actual color to be used is a bit tricky. The
palette has a number of colors through combinations of color group
and color role but none of those seem to actually be the focus rect
color. Thefore this is now hardcoded to some some shade of blue.