Set a flag to check if a node was actually used in move tool
Set a flag to check if a node was actually used in move tool. The
problem is that if snap is enabled snapping to a grid position on
mouse release moves the object even if it was not actually moved.
I.e. if the user just clicks on a node the move tool activates the
selected node is snapped to some location. Not exactly what was
wanted most likely.