Add first version of cut/copy/paste functionality
Add first version of cut/copy/paste functionality. While these
functions itself work well there are some gross UI problems.

We do not want to accidentally stomp onto any built-in Qt widget
cut/copy/paste behaviour. Therefore the cut/copy/paste should only
be enabled when our custom widget (the "main/cetral" widget inside
the MainWidget (most likely GfxWidget) has focus. I.e. the implementation
does not care any other cut/copy/paste functionality some other widget
implements and is not interested in forwarding actions to those
other focused widgets *except* for the QPlainTextEdit inside the
ScriptWidget.

This means that there are two ways to handle proper "disable" of
these actions when the keyboard focus moves to whatever widget in
the application:

1) track state changes and disable cut/copy/paste actions on the
   state changes.  I.e. handle widget focus change events. This is
   likely not even going to work because of the fact that the impl
   inside GfxWidget is a GfxWindow which is a QWindow and not a
   QWidget.

2) indicate the state on the action's in the menu when the menu
   is opened (i.e. query the widget whether the action can be done
   or not) and then *ignore* the action if invoked through the
   keyboard shortcut when the state is such that the action cannot
   be done. This is generally simpler but it means that when the
   paste action for example is invoked the MainWidget's paste impl
   needs to check that the GfxWidget has focus.

Now here's the problem: Because GfxWidget implementation is actually
a QWindow and not a QWidget the check for focus is broken because when
the menu opens the GfxWidget's QWindow loses the focus. Thus if the
paste action is invoked through the menu and the code does a check
for whether the GfxWidget/GfxWindow has focus this check will fail
and the paste action gets ignored. Seems that the only fix for this
would be to have the paste functionality *not* check for focus when
being invoked through the menu as the check whether paste is ok
has already been done before the menu opens.

Other problem is that the ScriptWidget delegates the queries for
wheter cut/copy/paste can be enabled to the QPlainTextEdit. The check
for canPaste seems to producing mostly false regardless of the check
whether the widget itself has focus or not. This means that the paste
action is then incorrectly disabled however the actual paste when done
through keyboard shortcut still works beacuse the key processing
is done in the QPlainTextEdit itself. In summary the core functionality
works but the UI state management is a clusterfuck.

Finally the copy/paste uses a simple text based custom clipboard that
is only available within this application instead of the global
QClipboard this is simply because he data interchange between
applications would not work anyway even if the data was placed onto
the system clipboard.