Fix an assert when adding a new workspace resource
Fix an assert when adding a new workspace resource. The filtering
proxy that was added in 6910d5c86f537abe791a4ec146fdd0ec930ff7f5
will be triggered when endInsertRows is called in the workspace.
This means the counter for user defined resources must be updated
before calling endInsertRows in order to properly account for the
new number of resources, otherwise we'll hit an incorrect assert
in GetUserDefinedResource.