Hide the complexity of traversing the scene nodes inside scene
Hide the complexity of traversing the scene nodes inside scene.
There are some differences in how the entities are transformed with
respect to scene class and the actual scene. When we're dealing with
scene class the problem is that the data for placing the entity
somewhere in the scene needs to exist somewhere but we cannot modify
the scene class itself since that's a resource that is shared between
all scenes. Therefore the SceneNodeClass needs to exist. However when
dealing with an actual instance of the scene and instances of
entities the placement data can be put inside each entity itself and
there is no need for any additional data holder. By moving the traversal
of the scene inside the scene it's possible to provide a simpler
interface to the consumers (such as the renderer) of the scene.
There'd still probably need to be some parameter that allows only
certain nodes to be collected. For example only collect the nodes
that are inside some rectangle.