Combine initial entity placement transform into entity root nodes
Combine initial entity placement transform into entity root nodes.

When the physics engine moves objects it's only looking at entity
nodes with rigid bodies. The entity nodes are then transformed
and their world transformation is expressed as a transformation
relative to each node's parent node.  This means that only the entity
nodes transforations would change but not the entity object's
transformation itself. This leads to silly problems such as not
really knowing where the entity is in the world.

Consider having an entity with some top level nodes. (not necessarily
a good design for the entity but still). The entity when placed
into the scene and linked to some other entity has an initial
transformation relative to it's linked parent entity (and the
selected node). If the  nodes have rigid bodies they will start to
move (without constraint joint to the parent node). And their node
transformations  will then change. If the  game code then proceeds
to for example ask "what's the position of the entity in the world"
the question is probably not what they really want to ask, but
rather they would want to ask "what is the position of some
particular entity node in the world". So in a sense the entity
transform is rather meaningless and superfluous only having meaning
for the initial transform state.

This problem can by solved by removing any such transformation
from the entity object itself and the initial transformation
can be baked directly into the top level entity nodes.