Improve the object selection algorithm in scene widget on mouse click
Improve the object selection algorithm in scene widget on mouse click.
Instead of simply just looking at the hit boxes of objects very coarsely
also try to discard items where the clickpoint is a pixel outside of
the actual shape. This can be done by rendering the objects returned
by the coarse hit test with different material colors and then doing
a readback in order to determine which object's pixel was actually
under the hit coordinate in the rendering surface.