Support basic undo in scene and entity widgets
Support basic undo in scene and entity widgets. The implementation
is the only "sane" way of doing undo that I could think of without
having to do a ton of work by creating a QUndoCommand for *every*
user action. Rather than make undo commands we simply take snapshots
of the whole state on timer intervals and store them in the  undo
buffer when there are changes. This isn't very sophisticated or
efficient but at least the implementation is simple.