Call base class (QWindow) keyPressEvent handler from GfxWindow
Call base class (QWindow) keyPressEvent handler from GfxWindow. The
Qt docs for QWindow don't mention anything about having to call the
base class if the re-implementation of the derived class's key press
handler doesn't act on the key *but* the QWidget documentation does
mention this. There has been some weird hicups with some keyboard
occasionally not working so I'm guessing that the Qt docs are simply
not up to date and in fact one should call QWindow::keyPressEvent
if not acting on the key.