Facelift the particle engine visualization in the editor widget
Facelift the particle engine visualization in the editor widget.
Option to display the regular background coordinate grid and also
to indicate the bounds of the visualization. Remove the x and y
translation parameters from the visualization transform since they
seem kinda pointless. Simply position the visualization around
the center of the window.