Remove rotation parameter from the simple gfx drawing functions
Remove rotation parameter from the simple gfx drawing functions. The
reason why it's removed is because it doesn't make that much sense
in terms of positioning the shapes. These API functions assume top
left origin which is used as the position for the shape. This doesn't
work well with rotation, as in rotation changes the position. To make
some sense the position would have to be the barycentric center of the
shape. But that API functionality is already covered elsewhere by
using the Transform object.