Add support for pixel ratio and pixel ratio based particle scaling
Add support for pixel ratio and pixel ratio based particle scaling.
The pixel ratio expresses the ratio of render units (pixels, texels)
to game units and depends on the "zoom" level of the current viewport.

Effects such as GL_POINTS based particle rendering requires the size
of each particle to expressed in pixels, therefore in order to correctly
scale the particles the ratio of game units and pixels needs to be
known.

This change also changes the particle size setting from pixels to
logical game units (which are only ever the same when the ratio of
pixels to game units is 1.)