Fix stale texture binding bug in gfx program implementation
Fix stale texture binding bug in gfx program implementation. The
bug is that when the same program object is used to do multiple draws
within one begin/end draw cycle the texture array in the program
implementation retains its old bindings from the previous draw.
This means that the the next draw could incorrectly bind also some
stale textures even though it doesn't want to use those.

The "fix" here is to add a new method in the gfx::Program interface
that is used to specify the number of textures the next draw will
use. This is a little big ugly and bug prone, a better solution would
be to have completely new API semantics where the texture binding
method takes an array/vector of texture binding data. Leaving this
as a todo for later now since the scope of the change is currently
minimal.