Add simple background rendering in the engine using a scene
Add simple background rendering in the engine using a scene. The
background is simply a scene object that is rendered before the
game scene which is placed on top (on the foreground). The background
scene is always placed in the middle of the viewport and does not
have physics simulation.