Add a WAR for not being able to load shaders during project play
Add a WAR for not being able to load shaders during project play.
The fundamental issue is still that the shader references in the
graphics subsystem are all "raw" i.e. not encoded in the resource
URI encoding, this means that they need to be resolved and treated
as file paths relative to the running application's working folder.
However this means that when the project play process is launched
the shaders are not necessarily deployed anywhere in the game dirs
(if no packaging has been done) and thus cannot be found. So this
commit adds a little WAR (more like a hack) to look for the shaders
in the the editor's (game host's) path.

Also this commit fixes the last remaining raw shader reference in
pinyin invaders. It now uses a fs:// locator and gets installed to
the dist folder during the project build since it's not being used
by any of the game assets (materials).