Refactor material actuator into single value interpolator
Refactor material actuator into single value interpolator. This
solves a bunch of problems related to the previous material actuator
that didn't even apply to a material in the first place. This
commit changes that into a value setter actuator similar to set flag
actuator expept that parameter to be changed is a float value and
is interpolated over time.