Remove the qdarkstyle which is based on a qss stylesheet
Remove the qdarkstyle which is based on a qss stylesheet. It makes
the theme setting more complicated since this stylesheet can only
be applied to Fusion(?) style. Since it's possible to use kvantum
on Windows we can delegate the "looking good" part completely to
a style engine. This should also simplify the process of having
uniform look when leveraging QtAssistant for the help system.