Re-layout and regroup the mainwidgets in the editor app
Re-layout and regroup the mainwidgets in the editor app. Remove the
top level group boxes since they're a little bit superfluous and
instead move the central gfx widget inside a viewport groupbox and
then regroup everything into a horizontal layout. Without the central
groupbox for the gfx widget the UI looks a little bit wrong as if
something was missing.