Refactor rasterizer state setting from drawable state
Refactor rasterizer state setting from drawable state. Currently this
state includes the line width setting when drawing lines and also
(as a new addition) polygon face culling setting. Instead of setting
this state into the drawable set the state to a rasterizer state object
in a new api function and then pass to the device in a device state
object.