Add an entity flag for flipping a drawable item about Y axis
Add an entity flag for flipping a drawable item about Y axis. This
can be convenient for example for turning the player character's
visual representation around to face another direction.

The flip cannot currently pertain to the whole of the entity since
flipping the physics body isn't allowed (need to double check) by
Box2D. I.e. once the physics body and the shapes attached to it
are created the shapes must remain constant with respect to the
bodies they're attached to. This means that flipping an entity with
a rigid body would only work correctly when the physics node shape
would be symmetrical.