Enhance the simple drawing API to provide a more generic API
Enhance the simple drawing API to provide a more generic API using
drawable's as the shape to be rasterized inside the bounding rect.

Todo: the current algorithm for drawing the outline of a shape doesn't
produce good results with diagonal lines when drawing the outline
of right triangle for example. This needs improvement.