Improve collision shape setting and add more collision shapes
Improve collision shape setting and add more collision shapes. Add
"primitive" collision shapes for some of the primitive drawable
shapes including: triangles, trapezoid and parallelogram. Unfortunately
it's not quite possible to exract the vertex data as such from the
drawable geometry types without restructing those since they also
deal with setting up the GPU vertex buffers and the draw commands.
These primitives are quite simple and we're only interested in the
convex hulls so the current implementation currently duplicates
the shape's geometry information in the physics system. Thus if the
shapes are altered in the graphics system the appropriate adjustments
need also be done in the physics representation.

Also add the (pre-)computation of a convex hull for a polygon shape.
The issue is that when extracting the vertices from a polygon shape
the shape can contain more vertices than what are in the convex hull.

It's unclear how Box2D behaves when the number of vertices exceeds
the b2_maxPolygonVertices. It seems that precomputing this convex
hull and discarding the non-interesting vertices imrpoves the physics.
For example the bandit demo with the stone block with slanted corners
begins to behave much better.