Disable keyboard grab in playwindow since it seems to work without
Disable keyboard grab in playwindow since it seems to work without.
When the playback is started and the play window comes up and the
keyboard input focus is where it's supposed to be, i.e on the
surface window. It maybe works now when the activation request is
done on timer. Who knows  ¯\_(ツ)_/¯