Delete Lua script file when the script resource object is deleted
Delete Lua script file when the script resource object is deleted.
Script resources are special in the sense that they're the only
resources where the underlying file system object is created by
the editor application. For everything else the content file on the
filesystem is created by some other application/tool and the workspace
only keeps a reference to that file. But since the script file is
(most likely) created by the editor we're just going to go ahead
and delete it. This could maybe use some confirmation from the
user but that's for later.