Add support for entity type specific Lua scripts
Add support for entity type specific Lua scripts. These are scripts
that are attached to the entity type created in the editor. When
game play is ongoing the script methods are called on every instance
of that particular type. This should allow one to write more
independent Lua scripts where functionality and behaviour related to
some particular type can be coded in its own specific script.

Refactor bandit demo's main game script so that the functionality
related to controlling the player entity are now in the player script.