Fix nullptr access on texture samplers that were optimized away
Fix nullptr access on texture samplers that were optimized away by
the shader compiler. The program object keeps a vector of texture
objects that are bound to the shader uniforms, but there's also
a check for the case when the glUniformLocation returns -1 i.e.
the texture uniform has been removed as an optimization. In this case
we skip adding the vector into the program's texture array which
later results in the texture object being nullptr.

Fix this by allowing nullptrs to exist in the texture vector and
add logical checks for this condition. This also means that the
loop for trying to bind the texture object to a texture unit can
be completely skipped.