Fix click point position check when selecting a tool in entity widget
Fix click point position check when selecting a tool in entity widget.
The tool corner box is defined in widget's units (i.e. pixels) so in
order to check for the tool box hit correctly the relative size of
the tool box  needs to be mapped into the node's model coordinate
by inverse scale.