Rename some entity and entity class get/find functions
Rename some entity and entity class get/find functions. This naming
should be more in line with the general naming philosophy. Get is
for things that aren't expected to fail or not be able to produce
the value while find is used for "looking up" when something may
not be there. Looking up the entity/node transforms may fail if the
nodes aren't found as expected.