Add missing file name conflict resolution when packing textures
Add missing file name conflict resolution when packing textures.
There's still a possibility for a filename collision to happen
when a texture is resized but this needs to be refactored separately.
In fact the concept of "texture size" needs to be split into 2
so that max texture size is independent from the max texture atlas
size. When this change is done the renaming can then also be done
for this case.