Add first support for adding audio/data file resources in workspace
Add first support for adding audio/data file resources in workspace.
These resources are not supported by the editor so the only purpose
to be able to add them to the resource tree is for being able to
pack them when the workspace resources are packaged.

This would still need further enhancements:
- File name conflict resolution and lookup for filenames.
- A way to invoke external tools on these files.
- An improved/alternative way to import/add them into into the
  workspace.

Refactor invaders to load the audio and data files from new changed
locations per packing output.