Refactor per entity node animation track saving and management
Refactor per entity node animation track saving and management. This
commit changes the way the per node animation tracks (timelines)
are managed. Previously they were only implied in the sense that
each entity node would implicitly map to one timeline for its
actuators. This has the problem that it will become difficult in
the UI to express overlapping actuators. I.e. when several actuators
should be running at the same time. Letting the actuators overalap
on a single timeline/track would make things rather messy. A better
solution is to have the possibility to have multiple timelines per
each entity node. This however means that this data now needst to
be saved and managed somehow separately of the actual animation or
actuator data since it only pertains to the UI. With this change
the data for the timelines is stored separately in the entity resource
and this makes it possible to have multiple per node animation tracks.