Update Editor starter content with ground box and blinking message
Update Editor starter content with ground box and blinking message.
The ground plane stops the bird from falling out of the view and
the message box demonstrates the use of a text item in an entity
node and gives the player a simple instruction on how to fly the
bird.