Add first version of UI Kit (uik for short)
Add first version of UI Kit (uik for short). This change implements
the first primitives for creating various game UIs. Below is a list
of the main classes/interfaces in this change.

Window - defines the high level container for widgets and provides
API for designing widget hieararchies and then dispatching events
such as mouse and keyboard events to the right widgets at runtime.

Painter - defines the abstract interface for painting widgets.
There's no actual implementation in order to keep the dependencies
to a minimum. Each widget will call the appropriate painter methods
from its Paint implementation in order to put some pixels on the
screen and it's up to the implementantion to provide some meaningful
way of doing it.

Widget - the main widget interface with which to interact with any
particular widget. Provides ways to interact with the widget both
at design and at runtime. Any particular widget implementation will
use the Painter interface to paint itself and provides an implementation
for the event handlers (mouse and keyboard) possibly translating
any such events into widget actions such as button presses or value
changes.

The widgets can be structured into a arbitrarily deep hierarchy of
parent-child widgets. Any window object will always have a root
widget (which is of type Form) into which all the other widgets in the
widget will be added. The algorithms that operate on the widget
hierarchy are then available in uikit/op.h as free standing functions
although for convenience for most operations the window will provide
a more straightforward wrapper function.

Currently only a few simple widget types are implemented:

Label      - Shows a simple string text and cannot be interacted with.
PushButton - Shows a button which can be clicked. On clicks ButtonPress
             actions are generated.
GroupBox   - Shows a group container with a widget border. Can contain
             other widgets. Cannot be interacted with.
CheckBox   - Shows a binary toggle box that is either on or off. On
             clicks will toggle state and generate ValueChanged events.
Form       - Shows a background form. Can contain other widgets.
             Cannot be interacted with.

Current list of todo items:

- More widget types such as progress bar, combo box, list widget.
- Keyboard handling API for passing keyboard events. This includes
  providing virtual keys for keyboard actions such as "select" or
  "focus next widget" or "focus previous widget".
- Window needs keyboard focus handling. It'll need to react to some
  special keys and pass others to the widget with the keyboard focus.
- Widgets need keyboard handling specific and applicable to each
  widget type.
- As more complicated widgets are created more primitive drawing api
  needs to be implemented. Such as drawing arrow down/up buttons for
  cases like combobox drop down button.
- More unit tests.