Fix graphics device texture filter setting change on change on default
Fix graphics device texture filter setting change on change on default.
If the device's default texture filtering settings are changed the
change is not properly applied on the texture unit. The fix is to
always map the texture's filtering setting to a device specific
setting and then check against the cached state. Now if the default
changes the unit's value will also get changed.