Refactor nlohmann::json utility code into base/json.h, base/json.cpp
Refactor nlohmann::json utility code into base/json.h, base/json.cpp.
The big problem with nlohmann::json is that including the json.hpp
file slows down the compilation considerably since this is a very
large template heavy header. This commit will do the initial lifting
of moving the utility functionality into a separate header and move
as much of the code as possible into a translation unit in order to
reduce the compilation effort. The following still needs to be done
in separate commits:

- Remove the templates from the header so that only a nlohmann::json
  forward declaration is needed.

- Remove other mentions of nlohmann/json.hpp in other parts of the
  code case by case.