Use only nlohmann/json_fwd.hpp in gfx/bitmap.h
Use only nlohmann/json_fwd.hpp in gfx/bitmap. This moves the bitmap
noise generator json related implementation inside a translation
unit in order to try to reduce the build drag.

Combine the gfx source files into a testing library that is used
when building the unit tests in order to reduce the build drag.