Try another shortcut key binding for game play
Try another shortcut key binding for game play. Working in the editor
is largely mouse driving. For a right handed person it's natural
to use left hand on the keyboard and drive mouse with the right hand
which means that hitting the Shift+F5 to play key combo is not that
easy. Trying F4 for now.