Remove background scene feature from game engine
Remove background scene feature from game engine. This is a messy
feature and makes things more complicated in the engine and brings
up questions such as "should this scene run physics?". Currently
it's possible to achieve the same effect (for example the parallax
scrolling in Bandit demo) by placing the background entity in the
"main" scene itself and moving it properly wrt the game viewport.
This can be easily achieved by for example adding the background
object to the player in the scene graph.