Add first support for game UI state management and action handling
Add first support for game UI state management and action handling.
This change adds new Lua side APIs for opening and closing game UIs.
Whenever there are uik::Window objects in the engine's UI stack the
topmost window object is given a chance to handle the input from the
mouse. If the mouse input results in some widget action the action
is passed to the game object for reacting on.