Refactor (game) time management into app implementations
Refactor (game) time management into app implementations. The loader
application still provides the time increments (and has to continue
to do that so that time increments can be paused in editor) but the
details are too various and possibly specific to a game so that it
makes sense to move them closer to the game logic and not the loader.

Remove the mention of "wall time" in the Lua api. It's confusing and
the wall time can be make large jumps if the game time is halted and
then unhalted (and updates are called again).