Refactor and refine runtime scene entity spawning and killing
Refactor and refine runtime scene entity spawning and killing. The
important things to conside here are:

- The lifetime of the C++ object must remain valid during the normal
  game play activities and while Lua code is invoked. In other words
  it should not be possible for the Lua game code to invalidate any
  game play objects (entities) it's currently playing with.

- Iteration of the current entities in the scene while invoking
  arbitrary game Lua code must be safe. For example when the game
  ticks the objects in the scene are iterated over and entity code
  in game's Lua code will be invoked for each entity. The Lua code
  can then in turn want to alter the scene by removing or by adding
  new game play objects (entities). That means that both operations
  must be safe to perform *while* the calling code is iterating over
  the current list of objects.

== Spawn ==

Spawning new entities happens in 2 stages.

Entities spawned by SpawnEntity are placed into a separate spawn list.
This will not invalidate any current iteration over the scene's entities.

On the next call to BeginLoop the entities on the spawn list are moved
into the actual list of entities in the scene and their Spawn control
flag is set. The spawn control flag is set only exactly for one
iteration of the game loop and will be cleared on the next call to
EndLoop. This allows any game sub-system to react to the spawning
of a new entity exactly once.

== Kill ==

Killing existing entities happens in 2 stages.

Entities killed by KillEntity are noted in a separate kill list and
not yet removed from the scene's entity list. This will not invalidate
any current iteration over the scene's entities.

On the next call to BeginLoop the entities on the kill list will have
their Kill control flag turned on. The next call to EndLoop will then
delete the entities with kill flags and remove them from the scene's
entity list. This allows any game subystem to react to the control
flag exactly only once.

This also means that any well-behaved game subsystem may no longer
use any pointer after the kill flag has been seen once since the
object is likely to be deleted.

TODO: It might be a good idea to let the Lua integration employ
reference counting regarding entity objects just to make sure that
if any Lua based game code tries to stash object references they'd
not become invalid/dangling pointers.